feat: add message transformation utility for AiSDK and update exports in ai-sdk module
