+++ packages/mcp/mcp-client-chat/src/ai/ai-sdk/index.ts
@@ -1,1 +1,2 @@
 export * from './ai-sdk.js';
+export * from './utils.js';
