+++ packages/mcp/mcp-client-chat/src/index.ts
@@ -2,3 +2,4 @@ export * from './init.js';
 export * from './types/index.js';
 export * from './re-act/index.js';
 export * from './function-calling/index.js';
+export * from './ai/index.js';
